requirements.txt
opencv-python
numpy
onnxruntime-gpu